修改字段映射

+++ healthcare/src/main/java/com/kanglian/healthcare/back/dal/pojo/UserInfo.java
@@ -1,14 +1,12 @@
 package com.kanglian.healthcare.back.dal.pojo;
 
 import java.util.Date;
-
 import com.easyway.business.framework.json.annotion.NotJsonData;
 import com.easyway.business.framework.pojo.BasePojo;
-import java.math.BigInteger;
 
 public class UserInfo extends BasePojo {
 	private static final long serialVersionUID = 1L;
-	private BigInteger id;
+	private Long id;
 	private Integer userId;
 	private String orgCode;
 	private String patientId;
@@ -26,10 +24,10 @@ public class UserInfo extends BasePojo {
 	private String cardNo;
 	private String drugAllergyMark;
 	private Date lastUpdateDtime;
-	public BigInteger getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(BigInteger id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public Integer getUserId() {
